revert github workflows to use java8, java11, java17
github workflow should use the various java versions to test,
which permits to test especially 2d rendering differences.

the javac release flag is set to produce java8 binaries in
any case. not perfect, but we only loose a warning that
applets are to be removed after java17, trivial to display
by adding the javac release property, when using java17:

  mvn --batch-mode compile -Dmaven.compiler.release=17
  gradle clean build -x javaDoc -PjavacRelease=17

better would be to separate build and test, as suggested
by https://github.com/matthew16550, here:
https://github.com/plantuml/plantuml/pull/904